+++ pyobs/modules/camera/basecamera.py
@@ -113,7 +113,7 @@ class BaseCamera(
         self._sequence_delay_abort = asyncio.Event()
 
         # register exception
-        exc.register_exception(exc.GrabImageError, 3, timespan=600, callback=self._default_remote_error_callback)
+        self._register_exception(exc.GrabImageError, 3, timespan=600, callback=self._default_remote_error_callback)
 
     async def open(self) -> None:
         """Open module."""