+++ pyobs/modules/flatfield/flatfield.py
@@ -90,15 +90,15 @@ class FlatField(Module, IFlatField, IBinning, IFilters):
 
         # register exceptions
         if isinstance(camera, str):
-            exc.register_exception(
+            self._register_exception(
                 exc.RemoteError, 3, timespan=600, module=camera, callback=self._default_remote_error_callback
             )
         if isinstance(telescope, str):
-            exc.register_exception(
+            self._register_exception(
                 exc.RemoteError, 3, timespan=600, module=telescope, callback=self._default_remote_error_callback
             )
         if isinstance(filters, str):
-            exc.register_exception(
+            self._register_exception(
                 exc.RemoteError, 3, timespan=600, module=filters, callback=self._default_remote_error_callback
             )