+++ pyobs/modules/focus/focusseries.py
@@ -79,11 +79,11 @@ class AutoFocusSeries(Module, CameraSettingsMixin, IAutoFocus):
 
         # register exceptions
         if isinstance(camera, str):
-            exc.register_exception(
+            self._register_exception(
                 exc.RemoteError, 3, timespan=600, module=camera, callback=self._default_remote_error_callback
             )
         if isinstance(focuser, str):
-            exc.register_exception(
+            self._register_exception(
                 exc.RemoteError, 3, timespan=600, module=focuser, callback=self._default_remote_error_callback
             )